+++ packages/react-native-gesture-handler/src/v3/hooks/callbacks/GestureHandlerStateChangeWorkletHandler.ts
@@ -1,7 +1,8 @@
 import { CALLBACK_TYPE } from '../../../handlers/gestures/gesture';
 import { State } from '../../../State';
 import { CallbackHandlers, StateChangeEvent } from '../../types';
-import { isEventForHandlerWithTag, isNativeEvent, runCallback } from '../utils';
+import { isEventForHandlerWithTag, isNativeEvent } from '../utils';
+import { runCallback } from '../utils/EventHandlersUtils';
 
 export function getGestureHandlerStateChangeWorkletHandler(
   handlerTag: number,
