+++ packages/react-native-gesture-handler/src/v3/hooks/callbacks/js/useGestureUpdateEvent.ts
@@ -1,6 +1,7 @@
-import { extractUpdateHandlers, isAnimatedEvent } from '../../utils';
+import { isAnimatedEvent } from '../../utils';
 import { ReanimatedContext } from '../../../../handlers/gestures/reanimatedWrapper';
 import { getGestureHandlerEventWorkletHandler } from '../GestureHandlerEventWorkletHandler';
+import { extractUpdateHandlers } from '../../utils/EventHandlersUtils';
 
 // eslint-disable-next-line @eslint-react/hooks-extra/ensure-custom-hooks-using-other-hooks
 export function useGestureUpdateEvent(handlerTag: number, config: any) {
