+++ packages/react-native-gesture-handler/src/v3/hooks/relations/useComposedGesture.ts
@@ -6,9 +6,9 @@ import {
   ComposedGesture,
   ComposedGestureType,
 } from '../../types';
-import { isComposedGesture } from '../utils';
 import { tagMessage } from '../../../utils';
 import { Reanimated } from '../../../handlers/gestures/reanimatedWrapper';
+import { isComposedGesture } from '../utils/RelationUtils';
 
 // TODO: Simplify repeated relations (Simultaneous with Simultaneous, Exclusive with Exclusive, etc.)
 export function useComposedGesture(
